@@ -1,6 +1,7 @@
 package com.termux.terminal;
 
 import android.util.Base64;
+import android.util.Log;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
@@ -28,65 +29,115 @@ import java.util.Stack;
  */
 public final class TerminalEmulator {
 
-    /** Log unknown or unimplemented escape sequences received from the shell process. */
+    /**
+     * Log unknown or unimplemented escape sequences received from the shell process.
+     */
     private static final boolean LOG_ESCAPE_SEQUENCES = false;
 
     public static final int MOUSE_LEFT_BUTTON = 0;
 
-    /** Mouse moving while having left mouse button pressed. */
+    /**
+     * Mouse moving while having left mouse button pressed.
+     */
     public static final int MOUSE_LEFT_BUTTON_MOVED = 32;
     public static final int MOUSE_WHEELUP_BUTTON = 64;
     public static final int MOUSE_WHEELDOWN_BUTTON = 65;
 
-    /** Used for invalid data - http://en.wikipedia.org/wiki/Replacement_character#Replacement_character */
+    /**
+     * Used for invalid data - http://en.wikipedia.org/wiki/Replacement_character#Replacement_character
+     */
     public static final int UNICODE_REPLACEMENT_CHAR = 0xFFFD;
 
-    /** Escape processing: Not currently in an escape sequence. */
+    /**
+     * Escape processing: Not currently in an escape sequence.
+     */
     private static final int ESC_NONE = 0;
-    /** Escape processing: Have seen an ESC character - proceed to {@link #doEsc(int)} */
+    /**
+     * Escape processing: Have seen an ESC character - proceed to {@link #doEsc(int)}
+     */
     private static final int ESC = 1;
-    /** Escape processing: Have seen ESC POUND */
+    /**
+     * Escape processing: Have seen ESC POUND
+     */
     private static final int ESC_POUND = 2;
-    /** Escape processing: Have seen ESC and a character-set-select ( char */
+    /**
+     * Escape processing: Have seen ESC and a character-set-select ( char
+     */
     private static final int ESC_SELECT_LEFT_PAREN = 3;
-    /** Escape processing: Have seen ESC and a character-set-select ) char */
+    /**
+     * Escape processing: Have seen ESC and a character-set-select ) char
+     */
     private static final int ESC_SELECT_RIGHT_PAREN = 4;
-    /** Escape processing: "ESC [" or CSI (Control Sequence Introducer). */
+    /**
+     * Escape processing: "ESC [" or CSI (Control Sequence Introducer).
+     */
     private static final int ESC_CSI = 6;
-    /** Escape processing: ESC [ ? */
+    /**
+     * Escape processing: ESC [ ?
+     */
     private static final int ESC_CSI_QUESTIONMARK = 7;
-    /** Escape processing: ESC [ $ */
+    /**
+     * Escape processing: ESC [ $
+     */
     private static final int ESC_CSI_DOLLAR = 8;
-    /** Escape processing: ESC % */
+    /**
+     * Escape processing: ESC %
+     */
     private static final int ESC_PERCENT = 9;
-    /** Escape processing: ESC ] (AKA OSC - Operating System Controls) */
+    /**
+     * Escape processing: ESC ] (AKA OSC - Operating System Controls)
+     */
     private static final int ESC_OSC = 10;
-    /** Escape processing: ESC ] (AKA OSC - Operating System Controls) ESC */
+    /**
+     * Escape processing: ESC ] (AKA OSC - Operating System Controls) ESC
+     */
     private static final int ESC_OSC_ESC = 11;
-    /** Escape processing: ESC [ > */
+    /**
+     * Escape processing: ESC [ >
+     */
     private static final int ESC_CSI_BIGGERTHAN = 12;
-    /** Escape procession: "ESC P" or Device Control String (DCS) */
+    /**
+     * Escape procession: "ESC P" or Device Control String (DCS)
+     */
     private static final int ESC_P = 13;
-    /** Escape processing: CSI > */
+    /**
+     * Escape processing: CSI >
+     */
     private static final int ESC_CSI_QUESTIONMARK_ARG_DOLLAR = 14;
-    /** Escape processing: CSI $ARGS ' ' */
+    /**
+     * Escape processing: CSI $ARGS ' '
+     */
     private static final int ESC_CSI_ARGS_SPACE = 15;
-    /** Escape processing: CSI $ARGS '*' */
+    /**
+     * Escape processing: CSI $ARGS '*'
+     */
     private static final int ESC_CSI_ARGS_ASTERIX = 16;
-    /** Escape processing: CSI " */
+    /**
+     * Escape processing: CSI "
+     */
     private static final int ESC_CSI_DOUBLE_QUOTE = 17;
-    /** Escape processing: CSI ' */
+    /**
+     * Escape processing: CSI '
+     */
     private static final int ESC_CSI_SINGLE_QUOTE = 18;
-    /** Escape processing: CSI ! */
+    /**
+     * Escape processing: CSI !
+     */
     private static final int ESC_CSI_EXCLAMATION = 19;
 
-    /** The number of parameter arguments. This name comes from the ANSI standard for terminal escape codes. */
+    /**
+     * The number of parameter arguments. This name comes from the ANSI standard for terminal escape codes.
+     */
     private static final int MAX_ESCAPE_PARAMETERS = 16;
 
-    /** Needs to be large enough to contain reasonable OSC 52 pastes. */
+    /**
+     * Needs to be large enough to contain reasonable OSC 52 pastes.
+     */
     private static final int MAX_OSC_STRING_LENGTH = 8192;
 
-    /** DECSET 1 - application cursor keys. */
+    /**
+     * DECSET 1 - application cursor keys.
+     */
     private static final int DECSET_BIT_APPLICATION_CURSOR_KEYS = 1;
     private static final int DECSET_BIT_REVERSE_VIDEO = 1 << 1;
     /**
@@ -104,40 +155,66 @@ public final class TerminalEmulator {
      * characters received when the cursor is at the right border of the page replace characters already on the page."
      */
     private static final int DECSET_BIT_AUTOWRAP = 1 << 3;
-    /** DECSET 25 - if the cursor should be enabled, {@link #isCursorEnabled()}. */
+    /**
+     * DECSET 25 - if the cursor should be enabled, {@link #isCursorEnabled()}.
+     */
     private static final int DECSET_BIT_CURSOR_ENABLED = 1 << 4;
     private static final int DECSET_BIT_APPLICATION_KEYPAD = 1 << 5;
-    /** DECSET 1000 - if to report mouse press&release events. */
+    /**
+     * DECSET 1000 - if to report mouse press&release events.
+     */
     private static final int DECSET_BIT_MOUSE_TRACKING_PRESS_RELEASE = 1 << 6;
-    /** DECSET 1002 - like 1000, but report moving mouse while pressed. */
+    /**
+     * DECSET 1002 - like 1000, but report moving mouse while pressed.
+     */
     private static final int DECSET_BIT_MOUSE_TRACKING_BUTTON_EVENT = 1 << 7;
-    /** DECSET 1004 - NOT implemented. */
+    /**
+     * DECSET 1004 - NOT implemented.
+     */
     private static final int DECSET_BIT_SEND_FOCUS_EVENTS = 1 << 8;
-    /** DECSET 1006 - SGR-like mouse protocol (the modern sane choice). */
+    /**
+     * DECSET 1006 - SGR-like mouse protocol (the modern sane choice).
+     */
     private static final int DECSET_BIT_MOUSE_PROTOCOL_SGR = 1 << 9;
-    /** DECSET 2004 - see {@link #paste(String)} */
+    /**
+     * DECSET 2004 - see {@link #paste(String)}
+     */
     private static final int DECSET_BIT_BRACKETED_PASTE_MODE = 1 << 10;
-    /** Toggled with DECLRMM - http://www.vt100.net/docs/vt510-rm/DECLRMM */
+    /**
+     * Toggled with DECLRMM - http://www.vt100.net/docs/vt510-rm/DECLRMM
+     */
     private static final int DECSET_BIT_LEFTRIGHT_MARGIN_MODE = 1 << 11;
-    /** Not really DECSET bit... - http://www.vt100.net/docs/vt510-rm/DECSACE */
+    /**
+     * Not really DECSET bit... - http://www.vt100.net/docs/vt510-rm/DECSACE
+     */
     private static final int DECSET_BIT_RECTANGULAR_CHANGEATTRIBUTE = 1 << 12;
 
 
     private String mTitle;
     private final Stack<String> mTitleStack = new Stack<>();
 
-    /** If processing first character of first parameter of {@link #ESC_CSI}. */
+    /**
+     * If processing first character of first parameter of {@link #ESC_CSI}.
+     */
     private boolean mIsCSIStart;
-    /** The last character processed of a parameter of {@link #ESC_CSI}. */
+    /**
+     * The last character processed of a parameter of {@link #ESC_CSI}.
+     */
     private Integer mLastCSIArg;
 
-    /** The cursor position. Between (0,0) and (mRows-1, mColumns-1). */
+    /**
+     * The cursor position. Between (0,0) and (mRows-1, mColumns-1).
+     */
     private int mCursorRow, mCursorCol;
 
-    /** The number of character rows and columns in the terminal screen. */
+    /**
+     * The number of character rows and columns in the terminal screen.
+     */
     public int mRows, mColumns;
 
-    /** The number of terminal transcript rows that can be scrolled back to. */
+    /**
+     * The number of terminal transcript rows that can be scrolled back to.
+     */
     public static final int TERMINAL_TRANSCRIPT_ROWS_MIN = 100;
     public static final int TERMINAL_TRANSCRIPT_ROWS_MAX = 50000;
     public static final int DEFAULT_TERMINAL_TRANSCRIPT_ROWS = 2000;
@@ -151,11 +228,15 @@ public final class TerminalEmulator {
     public static final int DEFAULT_TERMINAL_CURSOR_STYLE = TERMINAL_CURSOR_STYLE_BLOCK;
     public static final Integer[] TERMINAL_CURSOR_STYLES_LIST = new Integer[]{TERMINAL_CURSOR_STYLE_BLOCK, TERMINAL_CURSOR_STYLE_UNDERLINE, TERMINAL_CURSOR_STYLE_BAR};
 
-    /** The terminal cursor styles. */
+    /**
+     * The terminal cursor styles.
+     */
     private int mCursorStyle = DEFAULT_TERMINAL_CURSOR_STYLE;
 
 
-    /** The normal screen buffer. Stores the characters that appear on the screen of the emulated terminal. */
+    /**
+     * The normal screen buffer. Stores the characters that appear on the screen of the emulated terminal.
+     */
     private final TerminalBuffer mMainBuffer;
     /**
      * The alternate screen buffer, exactly as large as the display and contains no additional saved lines (so that when
@@ -164,20 +245,30 @@ public final class TerminalEmulator {
      * See http://www.xfree86.org/current/ctlseqs.html#The%20Alternate%20Screen%20Buffer
      */
     final TerminalBuffer mAltBuffer;
-    /** The current screen buffer, pointing at either {@link #mMainBuffer} or {@link #mAltBuffer}. */
+    /**
+     * The current screen buffer, pointing at either {@link #mMainBuffer} or {@link #mAltBuffer}.
+     */
     private TerminalBuffer mScreen;
 
-    /** The terminal session this emulator is bound to. */
+    /**
+     * The terminal session this emulator is bound to.
+     */
     private final TerminalOutput mSession;
 
     TerminalSessionClient mClient;
 
-    /** Keeps track of the current argument of the current escape sequence. Ranges from 0 to MAX_ESCAPE_PARAMETERS-1. */
+    /**
+     * Keeps track of the current argument of the current escape sequence. Ranges from 0 to MAX_ESCAPE_PARAMETERS-1.
+     */
     private int mArgIndex;
-    /** Holds the arguments of the current escape sequence. */
+    /**
+     * Holds the arguments of the current escape sequence.
+     */
     private final int[] mArgs = new int[MAX_ESCAPE_PARAMETERS];
 
-    /** Holds OSC and device control arguments, which can be strings. */
+    /**
+     * Holds OSC and device control arguments, which can be strings.
+     */
     private final StringBuilder mOSCOrDeviceControlArgs = new StringBuilder();
 
     /**
@@ -186,13 +277,17 @@ public final class TerminalEmulator {
      */
     private boolean mContinueSequence;
 
-    /** The current state of the escape sequence state machine. One of the ESC_* constants. */
+    /**
+     * The current state of the escape sequence state machine. One of the ESC_* constants.
+     */
     private int mEscapeState;
 
     private final SavedScreenState mSavedStateMain = new SavedScreenState();
     private final SavedScreenState mSavedStateAlt = new SavedScreenState();
 
-    /** http://www.vt100.net/docs/vt102-ug/table5-15.html */
+    /**
+     * http://www.vt100.net/docs/vt102-ug/table5-15.html
+     */
     private boolean mUseLineDrawingG0, mUseLineDrawingG1, mUseLineDrawingUsesG0 = true;
 
     /**
@@ -206,7 +301,9 @@ public final class TerminalEmulator {
      */
     private boolean mInsertMode;
 
-    /** An array of tab stops. mTabStop[i] is true if there is a tab stop set for column i. */
+    /**
+     * An array of tab stops. mTabStop[i] is true if there is a tab stop set for column i.
+     */
     private boolean[] mTabStop;
 
     /**
@@ -243,7 +340,9 @@ public final class TerminalEmulator {
      */
     int mForeColor, mBackColor;
 
-    /** Current {@link TextStyle} effect. */
+    /**
+     * Current {@link TextStyle} effect.
+     */
     private int mEffect;
 
     /**
@@ -252,7 +351,9 @@ public final class TerminalEmulator {
      */
     private int mScrollCounter = 0;
 
-    /** If automatic scrolling of terminal is disabled */
+    /**
+     * If automatic scrolling of terminal is disabled
+     */
     private boolean mAutoScrollDisabled;
 
     private byte mUtf8ToFollow, mUtf8Index;
@@ -414,22 +515,18 @@ public final class TerminalEmulator {
         return mCursorCol;
     }
 
-    /** Get the terminal cursor style. It will be one of {@link #TERMINAL_CURSOR_STYLES_LIST} */
+    /**
+     * Get the terminal cursor style. It will be one of {@link #TERMINAL_CURSOR_STYLES_LIST}
+     */
     public int getCursorStyle() {
         return mCursorStyle;
     }
 
-    /** Set the terminal cursor style. */
+    /**
+     * Set the terminal cursor style.
+     */
     public void setCursorStyle() {
-        Integer cursorStyle = null;
-
-        if (mClient != null)
-            cursorStyle = mClient.getTerminalCursorStyle();
-
-        if (cursorStyle == null || !Arrays.asList(TERMINAL_CURSOR_STYLES_LIST).contains(cursorStyle))
-            mCursorStyle = DEFAULT_TERMINAL_CURSOR_STYLE;
-        else
-            mCursorStyle = cursorStyle;
+        mCursorStyle = DEFAULT_TERMINAL_CURSOR_STYLE;
     }
 
     public boolean isReverseVideo() {
@@ -437,10 +534,10 @@ public final class TerminalEmulator {
     }
 
 
-
     public boolean isCursorEnabled() {
         return isDecsetInternalBitSet(DECSET_BIT_CURSOR_ENABLED);
     }
+
     public boolean shouldCursorBeVisible() {
         if (!isCursorEnabled())
             return false;
@@ -457,7 +554,6 @@ public final class TerminalEmulator {
     }
 
 
-
     public boolean isKeypadApplicationMode() {
         return isDecsetInternalBitSet(DECSET_BIT_APPLICATION_KEYPAD);
     }
@@ -466,7 +562,9 @@ public final class TerminalEmulator {
         return isDecsetInternalBitSet(DECSET_BIT_APPLICATION_CURSOR_KEYS);
     }
 
-    /** If mouse events are being sent as escape codes to the terminal. */
+    /**
+     * If mouse events are being sent as escape codes to the terminal.
+     */
     public boolean isMouseTrackingActive() {
         return isDecsetInternalBitSet(DECSET_BIT_MOUSE_TRACKING_PRESS_RELEASE) || isDecsetInternalBitSet(DECSET_BIT_MOUSE_TRACKING_BUTTON_EVENT);
     }
@@ -831,7 +929,7 @@ public final class TerminalEmulator {
                                 if (internalBit != -1) {
                                     value = isDecsetInternalBitSet(internalBit) ? 1 : 2; // 1=set, 2=reset.
                                 } else {
-                                    Logger.logError(mClient, LOG_TAG, "Got DECRQM for unrecognized private DEC mode=" + mode);
+                                    Log.e(LOG_TAG, "Got DECRQM for unrecognized private DEC mode=" + mode);
                                     value = 0; // 0=not recognized, 3=permanently set, 4=permanently reset
                                 }
                             }
@@ -886,7 +984,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** When in {@link #ESC_P} ("device control") sequence. */
+    /**
+     * When in {@link #ESC_P} ("device control") sequence.
+     */
     private void doDeviceControl(int b) {
         switch (b) {
             case (byte) '\\': // End of ESC \ string Terminator
@@ -975,7 +1075,7 @@ public final class TerminalEmulator {
                                     case "&8": // Undo key - ignore.
                                         break;
                                     default:
-                                        Logger.logWarn(mClient, LOG_TAG, "Unhandled termcap/terminfo name: '" + trans + "'");
+                                        Log.w(LOG_TAG, "Unhandled termcap/terminfo name: '" + trans + "'");
                                 }
                                 // Respond with invalid request:
                                 mSession.write("\033P0+r" + part + "\033\\");
@@ -987,12 +1087,12 @@ public final class TerminalEmulator {
                                 mSession.write("\033P1+r" + part + "=" + hexEncoded + "\033\\");
                             }
                         } else {
-                            Logger.logError(mClient, LOG_TAG, "Invalid device termcap/terminfo name of odd length: " + part);
+                            Log.e(LOG_TAG, "Invalid device termcap/terminfo name of odd length: " + part);
                         }
                     }
                 } else {
                     if (LOG_ESCAPE_SEQUENCES)
-                        Logger.logError(mClient, LOG_TAG, "Unrecognized device control string: " + dcs);
+                        Log.e(LOG_TAG, "Unrecognized device control string: " + dcs);
                 }
                 finishSequence();
             }
@@ -1015,7 +1115,9 @@ public final class TerminalEmulator {
         return mRightMargin - 1;
     }
 
-    /** Process byte while in the {@link #ESC_CSI_QUESTIONMARK} escape state. */
+    /**
+     * Process byte while in the {@link #ESC_CSI_QUESTIONMARK} escape state.
+     */
     private void doCsiQuestionMark(int b) {
         switch (b) {
             case 'J': // Selective erase in display (DECSED) - http://www.vt100.net/docs/vt510-rm/DECSED.
@@ -1082,7 +1184,7 @@ public final class TerminalEmulator {
                     int externalBit = mArgs[i];
                     int internalBit = mapDecSetBitToInternalBit(externalBit);
                     if (internalBit == -1) {
-                        Logger.logWarn(mClient, LOG_TAG, "Ignoring request to save/recall decset bit=" + externalBit);
+                        Log.w(LOG_TAG, "Ignoring request to save/recall decset bit=" + externalBit);
                     } else {
                         if (b == 's') {
                             mSavedDecSetFlags |= internalBit;
@@ -1272,7 +1374,7 @@ public final class TerminalEmulator {
                 // (1) enables this feature for keys except for those with well-known behavior, e.g., Tab, Backarrow and
                 // some special control character cases, e.g., Control-Space to make a NUL.
                 // (2) enables this feature for keys including the exceptions listed.
-                Logger.logError(mClient, LOG_TAG, "(ignored) CSI > MODIFY RESOURCE: " + getArg0(-1) + " to " + getArg1(-1));
+                Log.e(LOG_TAG, "(ignored) CSI > MODIFY RESOURCE: " + getArg0(-1) + " to " + getArg1(-1));
                 break;
             default:
                 parseArg(b);
@@ -1319,7 +1421,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** Encountering a character in the {@link #ESC} state. */
+    /**
+     * Encountering a character in the {@link #ESC} state.
+     */
     private void doEsc(int b) {
         switch (b) {
             case '#':
@@ -1412,7 +1516,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** DECSC save cursor - http://www.vt100.net/docs/vt510-rm/DECSC . See {@link #restoreCursor()}. */
+    /**
+     * DECSC save cursor - http://www.vt100.net/docs/vt510-rm/DECSC . See {@link #restoreCursor()}.
+     */
     private void saveCursor() {
         SavedScreenState state = (mScreen == mMainBuffer) ? mSavedStateMain : mSavedStateAlt;
         state.mSavedCursorRow = mCursorRow;
@@ -1426,7 +1532,9 @@ public final class TerminalEmulator {
         state.mUseLineDrawingUsesG0 = mUseLineDrawingUsesG0;
     }
 
-    /** DECRS restore cursor - http://www.vt100.net/docs/vt510-rm/DECRC. See {@link #saveCursor()}. */
+    /**
+     * DECRS restore cursor - http://www.vt100.net/docs/vt510-rm/DECRC. See {@link #saveCursor()}.
+     */
     private void restoreCursor() {
         SavedScreenState state = (mScreen == mMainBuffer) ? mSavedStateMain : mSavedStateAlt;
         setCursorRowCol(state.mSavedCursorRow, state.mSavedCursorCol);
@@ -1440,7 +1548,9 @@ public final class TerminalEmulator {
         mUseLineDrawingUsesG0 = state.mUseLineDrawingUsesG0;
     }
 
-    /** Following a CSI - Control Sequence Introducer, "\033[". {@link #ESC_CSI}. */
+    /**
+     * Following a CSI - Control Sequence Introducer, "\033[". {@link #ESC_CSI}.
+     */
     private void doCsi(int b) {
         switch (b) {
             case '!':
@@ -1761,7 +1871,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** Select Graphic Rendition (SGR) - see http://en.wikipedia.org/wiki/ANSI_escape_code#graphics. */
+    /**
+     * Select Graphic Rendition (SGR) - see http://en.wikipedia.org/wiki/ANSI_escape_code#graphics.
+     */
     private void selectGraphicRendition() {
         if (mArgIndex >= mArgs.length) mArgIndex = mArgs.length - 1;
         for (int i = 0; i <= mArgIndex; i++) {
@@ -1821,7 +1933,7 @@ public final class TerminalEmulator {
                 int firstArg = mArgs[i + 1];
                 if (firstArg == 2) {
                     if (i + 4 > mArgIndex) {
-                        Logger.logWarn(mClient, LOG_TAG, "Too few CSI" + code + ";2 RGB arguments");
+                        Log.w(LOG_TAG, "Too few CSI" + code + ";2 RGB arguments");
                     } else {
                         int red = mArgs[i + 2], green = mArgs[i + 3], blue = mArgs[i + 4];
                         if (red < 0 || green < 0 || blue < 0 || red > 255 || green > 255 || blue > 255) {
@@ -1846,7 +1958,7 @@ public final class TerminalEmulator {
                             mBackColor = color;
                         }
                     } else {
-                        if (LOG_ESCAPE_SEQUENCES) Logger.logWarn(mClient, LOG_TAG, "Invalid color index: " + color);
+                        if (LOG_ESCAPE_SEQUENCES) Log.w(LOG_TAG, "Invalid color index: " + color);
                     }
                 } else {
                     finishSequenceAndLogError("Invalid ISO-8613-3 SGR first argument: " + firstArg);
@@ -1863,7 +1975,7 @@ public final class TerminalEmulator {
                 mBackColor = code - 100 + 8;
             } else {
                 if (LOG_ESCAPE_SEQUENCES)
-                    Logger.logWarn(mClient, LOG_TAG, String.format("SGR unknown code %d", code));
+                    Log.w(LOG_TAG, String.format("SGR unknown code %d", code));
             }
         }
     }
@@ -1897,7 +2009,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** An Operating System Controls (OSC) Set Text Parameters. May come here from BEL or ST. */
+    /**
+     * An Operating System Controls (OSC) Set Text Parameters. May come here from BEL or ST.
+     */
     private void doOscSetTextParameters(String bellOrStringTerminator) {
         int value = -1;
         String textParameter = "";
@@ -1997,7 +2111,7 @@ public final class TerminalEmulator {
                     String clipboardText = new String(Base64.decode(textParameter.substring(startIndex), 0), StandardCharsets.UTF_8);
                     mSession.onCopyTextToClipboard(clipboardText);
                 } catch (Exception e) {
-                    Logger.logError(mClient, LOG_TAG, "OSC Manipulate selection, invalid string '" + textParameter + "");
+                    Log.e(LOG_TAG, "OSC Manipulate selection, invalid string '" + textParameter + "");
                 }
                 break;
             case 104:
@@ -2054,7 +2168,9 @@ public final class TerminalEmulator {
         return TextStyle.encode(mForeColor, mBackColor, mEffect);
     }
 
-    /** "CSI P_m h" for set or "CSI P_m l" for reset ANSI mode. */
+    /**
+     * "CSI P_m h" for set or "CSI P_m l" for reset ANSI mode.
+     */
     private void doSetMode(boolean newValue) {
         int modeBit = getArg0(0);
         switch (modeBit) {
@@ -2104,7 +2220,7 @@ public final class TerminalEmulator {
 
     /**
      * Process the next ASCII character of a parameter.
-     *
+     * <p>
      * Parameter characters modify the action or interpretation of the sequence. You can use up to
      * 16 parameters per sequence. You must use the ; character to separate parameters.
      * All parameters are unsigned, positive decimal integers, with the most significant
@@ -2112,16 +2228,16 @@ public final class TerminalEmulator {
      * (decimal). If you do not specify a value, a 0 value is assumed. A 0 value
      * or omitted parameter indicates a default value for the sequence. For most
      * sequences, the default value is 1.
-     *
+     * <p>
      * https://vt100.net/docs/vt510-rm/chapter4.html#S4.3.3
-     * */
+     */
     private void parseArg(int inputByte) {
         int[] bytes = new int[]{inputByte};
         // Only doing this for ESC_CSI and not for other ESC_CSI_* since they seem to be using their
         // own defaults with getArg*() calls, but there may be missed cases
         if (mEscapeState == ESC_CSI) {
             if ((mIsCSIStart && inputByte == ';') || // If sequence starts with a ; character, like \033[;m
-                (!mIsCSIStart && mLastCSIArg != null && mLastCSIArg == ';'  && inputByte == ';')) {  // If sequence contains sequential ; characters, like \033[;;m
+                (!mIsCSIStart && mLastCSIArg != null && mLastCSIArg == ';' && inputByte == ';')) {  // If sequence contains sequential ; characters, like \033[;;m
                 bytes = new int[]{'0', ';'}; // Assume 0 was passed
             }
         }
@@ -2222,7 +2338,7 @@ public final class TerminalEmulator {
     }
 
     private void finishSequenceAndLogError(String error) {
-        if (LOG_ESCAPE_SEQUENCES) Logger.logWarn(mClient, LOG_TAG, error);
+        if (LOG_ESCAPE_SEQUENCES) Log.w(LOG_TAG, error);
         finishSequence();
     }
 
@@ -2395,12 +2511,16 @@ public final class TerminalEmulator {
         mAboutToAutoWrap = false;
     }
 
-    /** Set the cursor mode, but limit it to margins if {@link #DECSET_BIT_ORIGIN_MODE} is enabled. */
+    /**
+     * Set the cursor mode, but limit it to margins if {@link #DECSET_BIT_ORIGIN_MODE} is enabled.
+     */
     private void setCursorColRespectingOriginMode(int col) {
         setCursorPosition(col, mCursorRow);
     }
 
-    /** TODO: Better name, distinguished from {@link #setCursorPosition(int, int)} by not regarding origin mode. */
+    /**
+     * TODO: Better name, distinguished from {@link #setCursorPosition(int, int)} by not regarding origin mode.
+     */
     private void setCursorRowCol(int row, int col) {
         mCursorRow = Math.max(0, Math.min(row, mRows - 1));
         mCursorCol = Math.max(0, Math.min(col, mColumns - 1));
@@ -2424,7 +2544,9 @@ public final class TerminalEmulator {
     }
 
 
-    /** Reset terminal state so user can interact with it regardless of present state. */
+    /**
+     * Reset terminal state so user can interact with it regardless of present state.
+     */
     public void reset() {
         setCursorStyle();
         mArgIndex = 0;
@@ -2461,12 +2583,16 @@ public final class TerminalEmulator {
         return mScreen.getSelectedText(x1, y1, x2, y2);
     }
 
-    /** Get the terminal session's title (null if not set). */
+    /**
+     * Get the terminal session's title (null if not set).
+     */
     public String getTitle() {
         return mTitle;
     }
 
-    /** Change the terminal session's title. */
+    /**
+     * Change the terminal session's title.
+     */
     private void setTitle(String newTitle) {
         String oldTitle = mTitle;
         mTitle = newTitle;
@@ -2475,7 +2601,9 @@ public final class TerminalEmulator {
         }
     }
 
-    /** If DECSET 2004 is set, prefix paste with "\033[200~" and suffix with "\033[201~". */
+    /**
+     * If DECSET 2004 is set, prefix paste with "\033[200~" and suffix with "\033[201~".
+     */
     public void paste(String text) {
         // First: Always remove escape key and C1 control characters [0x80,0x9F]:
         text = text.replaceAll("(\u001B|[\u0080-\u009F])", "");
@@ -2489,9 +2617,13 @@ public final class TerminalEmulator {
         if (bracketed) mSession.write("\033[201~");
     }
 
-    /** http://www.vt100.net/docs/vt510-rm/DECSC */
+    /**
+     * http://www.vt100.net/docs/vt510-rm/DECSC
+     */
     static final class SavedScreenState {
-        /** Saved state of the cursor position, Used to implement the save/restore cursor position escape sequences. */
+        /**
+         * Saved state of the cursor position, Used to implement the save/restore cursor position escape sequences.
+         */
         int mSavedCursorRow, mSavedCursorCol;
         int mSavedEffect, mSavedForeColor, mSavedBackColor;
         int mSavedDecFlags;
